add logging, add types mostly centered around Gamepads/Keyboards
There's a lot of stuff changed here but it mostly focused on trying
to get types going in ControlsModal, GamepadController, and
KeyboardController.

In the process, I added pino logging so that I could trace some of the
stuff that the application was doing and track down what the data was in
some of the data structures.